+++ leshan-client-core/src/test/java/org/eclipse/leshan/client/datacollector/ManualDataSenderTest.java
@@ -11,6 +11,9 @@ import java.util.Random;
 import java.util.stream.Collectors;
 
 import org.eclipse.leshan.client.datacollector.ManualDataSenderTest.FakeDataSenderManager.SendDataOutcome;
+import org.eclipse.leshan.client.send.DataSender;
+import org.eclipse.leshan.client.send.DataSenderManager;
+import org.eclipse.leshan.client.send.ManualDataSender;
 import org.eclipse.leshan.client.servers.ServerIdentity;
 import org.eclipse.leshan.core.node.LwM2mNode;
 import org.eclipse.leshan.core.node.LwM2mPath;
